server: refactor engine and server.
Add distributed_engine to call storage_engine.
Move the struct "Server" to "server.rs".
